Big code refactoring
* rename files and variables
* delete duplicated error types and use built in ones
* update and add more descriptions to cli tx cfeev commands
* simplify kv store keys
* instead of string type use proto's built-in types
* change proto int32 types to uint64 types (these fields cannot have negative values)
* update cfeev module codec
* add more logs
* general code refactoring
* add msg update denom to cfeev
* delete legacy params

+++ testutil/module/cfeev/cfeev_context_utils.go
@@ -14,7 +14,7 @@ func (h *ContextC4eEvUtils) VerifyEnergyTransferStatus(id uint64, status types.T
 	h.C4eEvUtils.VerifyEnergyTransferStatus(h.testContext.GetContext(), id, status)
 }
 
-func (h *ContextC4eEvUtils) EnergyTransferCompleted(energyTransferId uint64, usedServiceUnits int32) {
+func (h *ContextC4eEvUtils) EnergyTransferCompleted(energyTransferId uint64, usedServiceUnits uint64) {
 	h.C4eEvUtils.EnergyTransferCompleted(h.testContext.GetContext(), energyTransferId, usedServiceUnits)
 }
